requirements.txt
psycopg2-binary>=2.9.9
pandas>=2.1.0
openpyxl>=3.1.2
jinja2>=3.1.3
pyyaml>=6.0.1
tqdm>=4.66.0
